# docs/contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: DataX Enterprise Studio API
  version: 1.2.0
  license:
    name: License pending repository-owner decision
    url: https://github.com/xiaoli2hust/datax/blob/main/README.md
  description: |
    V1 单组织、多项目、安全的一次性离线全量复制控制 API。
    仅支持 MySQL 8/PostgreSQL 15、人工触发、源端静默确认、预存在且实测为空的目标表和 insert-only-once。
    不提供调度、DAG、告警、AI、CDC、自由 SQL、Transformer 或插件安装接口。
    交付形态为 Windows 11 x64 本地工作站；业务 API 仅经 web 反向代理绑定本机
    127.0.0.1:17860，api、worker 和 postgres 不暴露宿主端口。
servers:
  - url: http://127.0.0.1:17860/api/v1
    description: Windows 本地工作站唯一支持入口；不得替换为 0.0.0.0、LAN IP 或公网地址
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: 本地账号登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: operator@example.com
              password: <write-only-password>
      responses:
        '200':
          description: 登录成功；同时设置 HttpOnly refresh cookie
          headers:
            Set-Cookie:
              schema:
                type: string
              description: des_refresh；HttpOnly、SameSite=Strict、Path=/api/v1/auth、无 Domain。V1 loopback HTTP 不虚假声明 Secure。
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshAccessToken
      summary: 旋转 refresh token 并签发 access token
      security:
        - refreshCookie: []
      responses:
        '200':
          description: 刷新成功
          headers:
            Set-Cookie:
              schema:
                type: string
              description: 旋转后的 des_refresh；属性与登录响应一致
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: 幂等撤销调用方持有的本机会话凭证
      description: |
        接受 bearer、refresh cookie 或两者。每个能被密码学验证并精确定位的 session 都会
        被撤销；两者指向不同 session 时分别撤销，避免只删除 cookie 而留下服务器端 refresh
        session。缺失、无效、过期或已撤销凭证同样返回 204，响应始终删除本机 cookie。
      security:
        - bearerAuth: []
        - refreshCookie: []
        - {}
      responses:
        '204':
          description: 已撤销；重复退出同样返回 204
          headers:
            Set-Cookie:
              schema:
                type: string
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
  /auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: 获取当前用户及作用域角色
      responses:
        '200':
          description: 当前用户
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/change-password:
    post:
      tags: [Auth]
      operationId: changeOwnPassword
      summary: 修改本人密码并撤销其他 session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: 密码已修改；当前 session 保留，其他 session 已撤销
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: 分页列出组织用户
      description: 需要组织级 Admin。
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, LOCKED, DISABLED]
      responses:
        '200':
          description: 用户页
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Users]
      operationId: createUser
      summary: 创建本地用户
      description: 需要 Admin；temporary_password 是 write-only，响应不回显。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: 已创建，用户首次登录必须改密
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /users/{user_id}:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [Users]
      operationId: getUser
      summary: 获取用户状态与角色摘要
      description: 需要组织级 Admin。
      responses:
        '200':
          description: 用户
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Users]
      operationId: updateUser
      summary: 修改显示名或启用/停用用户
      description: 停用用户会撤销其全部 session。
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPatch'
      responses:
        '200':
          description: 已更新
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /users/{user_id}/unlock:
    parameters:
      - $ref: '#/components/parameters/UserId'
    post:
      tags: [Users]
      operationId: unlockUser
      summary: 解除登录锁定
      description: 需要 Admin；清零失败计数并清空 locked_until。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: 已解锁
          headers:
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /users/{user_id}/organization-roles:
    parameters:
      - $ref: '#/components/parameters/UserId'
    put:
      tags: [Users]
      operationId: replaceOrganizationRoles
      summary: 原子替换组织级 Admin 角色
      description: 需要 Admin；允许空数组撤销，但系统必须至少保留一名有效 Admin。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationRoleReplaceRequest'
      responses:
        '200':
          description: 已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /users/{user_id}/reset-password:
    parameters:
      - $ref: '#/components/parameters/UserId'
    post:
      tags: [Users]
      operationId: resetUserPassword
      summary: 设置临时密码并撤销全部 session
      description: 需要 Admin；响应不返回临时密码，用户下次登录必须改密。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '204':
          description: 已重置
          headers:
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: des_refresh
      description: V1 loopback HTTP 使用 HttpOnly、SameSite=Strict、Path=/api/v1/auth 且无 Domain；若未来改为受信 loopback HTTPS 必须增加 Secure。
  headers:
    RequestId:
      description: 服务端请求关联 UUID
      schema:
        type: string
        format: uuid
    ETag:
      description: 弱 ETag，值为 W/"row_version"
      schema:
        type: string
        pattern: '^W/"[1-9][0-9]*"$'
    IdempotencyReplayed:
      description: true 表示返回 24 小时内首次请求结果
      schema:
        type: boolean
  parameters:
...
    UserId:
      name: user_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      required: false
      description: 服务端生成的不透明游标
      schema:
        type: string
        minLength: 1
        maxLength: 2048
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
        pattern: '^W/"[1-9][0-9]*"$'
  responses:
...
    Unauthorized:
      description: 未认证、凭证无效或 token 到期
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: 无权限
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: 资源不存在或对当前主体不可见
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: 版本、幂等、状态或唯一约束冲突
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ValidationError:
      description: 请求或业务语义校验失败
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    Uuid:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
...
    Role:
      type: string
      enum: [ADMIN, DEVELOPER, OPERATOR, VIEWER]
...
    FieldError:
      type: object
      additionalProperties: false
      required: [path, code, message]
      properties:
        path:
          type: string
          maxLength: 500
        code:
          type: string
          pattern: '^[A-Z0-9_]+$'
        message:
          type: string
          maxLength: 1000
    CapacityAdmissionDetails:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason:
          type: string
          enum: [QUEUE_LIMIT, BACKLOG_LIMIT, DISK_YELLOW, DISK_RED, LOG_BUDGET]
    Problem:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - status
        - code
        - detail
        - instance
        - request_id
        - retryable
        - field_errors
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
          maxLength: 200
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          pattern: '^[A-Z0-9_]+$'
        detail:
          type: [string, 'null']
          maxLength: 2000
        instance:
          type: string
          maxLength: 500
        request_id:
          $ref: '#/components/schemas/Uuid'
        retryable:
          type: boolean
        field_errors:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/FieldError'
        details:
          oneOf:
            - $ref: '#/components/schemas/CapacityAdmissionDetails'
            - type: 'null'
...
    UserSummary:
      type: object
      additionalProperties: false
      required: [id, email, display_name, must_change_password]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        email:
          type: string
          format: email
          maxLength: 254
        display_name:
          type: string
          minLength: 1
          maxLength: 128
        must_change_password:
          type: boolean
          description: true 时除 refresh/logout/me/change-password 外的业务接口统一返回 PASSWORD_CHANGE_REQUIRED。
    User:
      type: object
      additionalProperties: false
      required:
        - id
        - email
        - display_name
        - status
        - must_change_password
        - role_assignments
        - row_version
        - created_at
        - updated_at
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        email:
          type: string
          format: email
          maxLength: 254
        display_name:
          type: string
          minLength: 1
          maxLength: 128
        status:
          type: string
          enum: [ACTIVE, LOCKED, DISABLED]
        must_change_password:
          type: boolean
        role_assignments:
          type: array
          items:
            $ref: '#/components/schemas/ScopedRoles'
        row_version:
          type: integer
          minimum: 1
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
    UserCreate:
      type: object
      additionalProperties: false
      required: [email, display_name, temporary_password]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        display_name:
          type: string
          minLength: 1
          maxLength: 128
        temporary_password:
          type: string
          format: password
          minLength: 12
          maxLength: 256
          writeOnly: true
    UserPatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        display_name:
          type: string
          minLength: 1
          maxLength: 128
        status:
          type: string
          enum: [ACTIVE, DISABLED]
    UserPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor, has_more]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/User'
        next_cursor:
          type: [string, 'null']
        has_more:
          type: boolean
    ChangePasswordRequest:
      type: object
      additionalProperties: false
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
          format: password
          minLength: 1
          maxLength: 256
          writeOnly: true
        new_password:
          type: string
          format: password
          minLength: 12
          maxLength: 256
          writeOnly: true
    ResetPasswordRequest:
      type: object
      additionalProperties: false
      required: [temporary_password]
      properties:
        temporary_password:
          type: string
          format: password
          minLength: 12
          maxLength: 256
          writeOnly: true
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          format: password
          minLength: 12
          maxLength: 256
          writeOnly: true
    AuthResponse:
      type: object
      additionalProperties: false
      required: [access_token, token_type, expires_in, user]
      properties:
        access_token:
          type: string
          readOnly: true
        token_type:
          const: Bearer
        expires_in:
          type: integer
          minimum: 60
          example: 900
        user:
          $ref: '#/components/schemas/UserSummary'
    ScopedRoles:
      type: object
      additionalProperties: false
      required: [scope_type, scope_id, roles]
      properties:
        scope_type:
          type: string
          enum: [ORGANIZATION, PROJECT]
        scope_id:
          $ref: '#/components/schemas/Uuid'
        roles:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Role'
    MeResponse:
      type: object
      additionalProperties: false
      required: [user, role_assignments]
      properties:
        user:
          $ref: '#/components/schemas/UserSummary'
        role_assignments:
          type: array
          items:
            $ref: '#/components/schemas/ScopedRoles'
...
    OrganizationRoleReplaceRequest:
      type: object
      additionalProperties: false
      required: [roles]
      properties:
        roles:
          type: array
          maxItems: 1
          uniqueItems: true
          items:
            const: ADMIN
...